+++ ReleaseNotes.md
@@ -3,3 +3,4 @@
 ## 1.2.0
 * symbol packages retain the symbols extension when updating the name
 * Added nuspec frameworkassemblies add command
+* Added files copysymbols command for merging pdb files